Added trimmed armor item datagen

+++ src/main/java/com/temporal/api/core/engine/io/metadata/annotation/data/other/TrimMaterialDescription.java
@@ -10,5 +10,5 @@ import java.lang.annotation.Target;
 public @interface TrimMaterialDescription {
     String itemId();
     String color();
-    int itemModelIndex();
+    float itemModelIndex();
 }
