+++ src/main/java/com/temporal/api/core/event/data/trim/material/TrimMaterialDescriptionHolder.java
@@ -1,4 +1,4 @@
 package com.temporal.api.core.event.data.trim.material;
 
-public record TrimMaterialDescriptionHolder(String itemId, String color, int itemModelIndex) {
+public record TrimMaterialDescriptionHolder(String itemId, String color, float itemModelIndex) {
 }
